+++ llvm/docs/ReleaseNotes.rst
@@ -44,6 +44,9 @@ Non-comprehensive list of changes in this release
 
 * The C API function LLVMAddTargetData has been removed.
 
+* The C API function LLVMGetDataLayout is deprecated
+  in favor of LLVMGetDataLayoutStr.
+
 .. NOTE
    For small 1-3 sentence descriptions, just add an entry at the end of
    this list. If your description won't fit comfortably in one bullet